+++ src/commonMain/kotlin/csense/kotlin/extensions/collections/Iterator.kt
@@ -53,3 +53,14 @@ public inline fun <T> Iterator<T>.toMutableList(
         list += it
     }
 }
+
+
+/**
+ * If this iterator is at the end (no more elements)
+ * @receiver Iterator<T>
+ * @see Iterator.hasNext (this is the inverse of that)
+ * @return [Boolean] true if there are no more elements. false if there are more elements.
+ */
+public inline fun <T> Iterator<T>.atEnd(): Boolean {
+    return !hasNext()
+}